Code Duplication Removal, Implemented Lombok for some of the model classes.
