+++ src/main/java/org/recap/controller/DeAccessionRestController.java
@@ -21,42 +21,9 @@ import java.util.Map;
  */
 @RestController
 @RequestMapping("/deAccessionService")
-public class DeAccessionRestController {
+public class DeAccessionRestController extends AbstractController{
 
     private static final Logger logger = LoggerFactory.getLogger(DeAccessionRestController.class);
-
-    @Value("${scsb.circ.url}")
-    private String scsbCircUrl;
-
-    @Autowired
-    private RestHeaderService restHeaderService;
-
-    /**
-     * Gets rest template.
-     *
-     * @return the rest template
-     */
-    public RestTemplate getRestTemplate() {
-        return new RestTemplate();
-    }
-
-    /**
-     * Gets scsb circ url.
-     *
-     * @return the scsb circ url
-     */
-    public String getScsbCircUrl() {
-        return scsbCircUrl;
-    }
-
-    /**
-     * Get rest header service rest header service.
-     *
-     * @return the rest header service
-     */
-    public RestHeaderService getRestHeaderService(){
-        return restHeaderService;
-    }
 
     /**
      * This method will call scsb-circ microservice to deaccession an item in SCSB.
