+++ src/main/java/org/recap/model/AbstractResponseItem.java
@@ -1,6 +1,5 @@
 package org.recap.model;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,11 +8,14 @@ import lombok.Setter;
  */
 @Getter
 @Setter
-public abstract class AbstractResponseItem {
+public abstract class AbstractResponseItem  {
+
     private String itemBarcode;
     private String itemOwningInstitution=""; // PUL, CUL, NYPL
     private String screenMessage;
     private boolean success;
     private String esipDataIn;
     private String esipDataOut;
+    private String bibId;
+
 }
