+++ src/main/java/org/recap/model/transfer/BaseSourceDestination.java
@@ -1,0 +1,11 @@
+package org.recap.model.transfer;
+
+import lombok.Data;
+
+@Data
+public class BaseSourceDestination {
+
+    private String owningInstitutionBibId;
+    private String owningInstitutionHoldingsId;
+
+}
