+++ src/main/java/org/recap/model/transfer/Destination.java
@@ -3,23 +3,6 @@ package org.recap.model.transfer;
 /**
  * Created by sheiks on 12/07/17.
  */
-public class Destination {
-    private String owningInstitutionBibId;
-    private String owningInstitutionHoldingsId;
+public class Destination extends BaseSourceDestination {
 
-    public String getOwningInstitutionBibId() {
-        return owningInstitutionBibId;
-    }
-
-    public void setOwningInstitutionBibId(String owningInstitutionBibId) {
-        this.owningInstitutionBibId = owningInstitutionBibId;
-    }
-
-    public String getOwningInstitutionHoldingsId() {
-        return owningInstitutionHoldingsId;
-    }
-
-    public void setOwningInstitutionHoldingsId(String owningInstitutionHoldingsId) {
-        this.owningInstitutionHoldingsId = owningInstitutionHoldingsId;
-    }
 }
