+++ src/main/java/org/recap/model/transfer/HoldingTransferResponse.java
@@ -1,8 +1,13 @@
 package org.recap.model.transfer;
 
+import lombok.Getter;
+import lombok.Setter;
+
 /**
  * Created by sheiks on 12/07/17.
  */
+@Setter
+@Getter
 public class HoldingTransferResponse {
     private String message;
     private HoldingsTransferRequest holdingsTransferRequest;
@@ -12,22 +17,6 @@ public class HoldingTransferResponse {
 
     public HoldingTransferResponse(String message, HoldingsTransferRequest holdingsTransferRequest) {
         this.message = message;
-        this.holdingsTransferRequest = holdingsTransferRequest;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public HoldingsTransferRequest getHoldingsTransferRequest() {
-        return holdingsTransferRequest;
-    }
-
-    public void setHoldingsTransferRequest(HoldingsTransferRequest holdingsTransferRequest) {
         this.holdingsTransferRequest = holdingsTransferRequest;
     }
 }
