+++ src/main/java/org/recap/model/transfer/HoldingsTransferRequest.java
@@ -1,25 +1,14 @@
 package org.recap.model.transfer;
 
+import lombok.Getter;
+import lombok.Setter;
+
 /**
  * Created by sheiks on 12/07/17.
  */
+@Getter
+@Setter
 public class HoldingsTransferRequest {
     private Source source;
     private Destination destination;
-
-    public Source getSource() {
-        return source;
-    }
-
-    public void setSource(Source source) {
-        this.source = source;
-    }
-
-    public Destination getDestination() {
-        return destination;
-    }
-
-    public void setDestination(Destination destination) {
-        this.destination = destination;
-    }
 }
