+++ src/main/java/org/recap/model/transfer/ItemSource.java
@@ -1,16 +1,13 @@
 package org.recap.model.transfer;
 
+import lombok.Getter;
+import lombok.Setter;
+
 /**
  * Created by sheiks on 13/07/17.
  */
+@Getter
+@Setter
 public class ItemSource extends Source {
     private String owningInstitutionItemId;
-
-    public String getOwningInstitutionItemId() {
-        return owningInstitutionItemId;
-    }
-
-    public void setOwningInstitutionItemId(String owningInstitutionItemId) {
-        this.owningInstitutionItemId = owningInstitutionItemId;
-    }
 }
