+++ src/main/java/org/recap/model/transfer/ItemTransferRequest.java
@@ -1,25 +1,14 @@
 package org.recap.model.transfer;
 
+import lombok.Getter;
+import lombok.Setter;
+
 /**
  * Created by sheiks on 13/07/17.
  */
+@Getter
+@Setter
 public class ItemTransferRequest {
     private ItemSource source;
     private ItemDestination destination;
-
-    public ItemSource getSource() {
-        return source;
-    }
-
-    public void setSource(ItemSource source) {
-        this.source = source;
-    }
-
-    public ItemDestination getDestination() {
-        return destination;
-    }
-
-    public void setDestination(ItemDestination destination) {
-        this.destination = destination;
-    }
 }
