+++ src/main/java/org/recap/model/transfer/ItemTransferResponse.java
@@ -1,8 +1,13 @@
 package org.recap.model.transfer;
 
+import lombok.Getter;
+import lombok.Setter;
+
 /**
  * Created by sheiks on 13/07/17.
  */
+@Getter
+@Setter
 public class ItemTransferResponse {
     private String message;
     private ItemTransferRequest itemTransferRequest;
@@ -12,22 +17,6 @@ public class ItemTransferResponse {
 
     public ItemTransferResponse(String message, ItemTransferRequest itemTransferRequest) {
         this.message = message;
-        this.itemTransferRequest = itemTransferRequest;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public ItemTransferRequest getItemTransferRequest() {
-        return itemTransferRequest;
-    }
-
-    public void setItemTransferRequest(ItemTransferRequest itemTransferRequest) {
         this.itemTransferRequest = itemTransferRequest;
     }
 }
